- Implemented Reader, Writer, StreamReader, StreamWriter (note that Writer.cpp is not part of the build but are notes for a couple of functions that may or may not be added to the Writer interface) - Added MarkSupported, Mark, and Reset to InputStream - Added OutputStream::WriteString - InputStream and OutputStream now use size_t instead of uint32_t for sizes and positions - on Windows, asserts will fire if the size or offset would overflow a DWORD - Removed GlobalSettings from System (Note: GlobalSettings is deprecated and will be removed) - Removed Win32 build config from VideoPlayer

+++ drivers/windows/io/WinFileOutputStream.h
@@ -41,7 +41,7 @@ namespace ee
 		virtual size_t GetCurrentOffset( void ) override final;
 
 		// Returns the number of bytes written
-		virtual uint32_t Write( const void* buffer, uint32_t length ) override final;
+		virtual uint32_t Write( const void* buffer, size_t length ) override final;
 		virtual void Flush( void ) override final;
 
 	private:
